Location inspecting tests now working for watin
One test set the location by script, but wasn't waiting for IE to be
ready before getting the location. The other failed in Watin +
WebDriver for IE because the test website was returning 404. When IE
receives the 404 and friendly error pages are enabled, it sets the
current location to
res://ieframe.dll/http_404.html#http://localhost:4567/
rather than just http://localhost:4567/

I've worked around the issue by getting the Sinatra site to have a
valid page at "/". But should we strip the friendly error page gumf?