Check FindElement for Displayed before fall back to FindElements
Since webdriver cannot give you just the visible elements the
original strategy was to use FindElements(by) then iterate until
one is displayed. Much faster to try FindElement(by) first and
see if that on is displayed, then fall back to the original
strategy. Big optimisation for the common case.