Performance issue with IE 64 bit driver.
Fillin method is super slow when using IE 64 bit driver.
The reason for this seem to be that the Fillin method is using selenium's sendkeys method. This method is for some reason very slow when using the 64 bit driver. Going from ca 5 sek pr char to almost instant. However the SeleniumBackedWebDriver have a method called Type(). This method runs fast even with the 64 bit driver for some reason.